+++ cloud-minecraft/cloud-bungee/src/main/java/cloud/commandframework/bungee/arguments/PlayerArgument.java
@@ -139,8 +139,8 @@ public final class PlayerArgument<C> extends CommandArgument<C, ProxiedPlayer> {
 
         @Override
         public @NonNull ArgumentParseResult<@NonNull ProxiedPlayer> parse(
-                @NonNull final CommandContext<@NonNull C> commandContext,
-                @NonNull final Queue<@NonNull String> inputQueue
+                final @NonNull CommandContext<@NonNull C> commandContext,
+                final @NonNull Queue<@NonNull String> inputQueue
         ) {
             final String input = inputQueue.peek();
             if (input == null) {