+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Added `@Suggestions` annotated methods
+ - Type safe meta system
 
 ### Changed
  - Argument builders now use generics to preserve their type when using functions from the base class.
@@ -16,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - Moved the parser injector registry into CommandManager and added injection to CommandContext
+
+### Deprecated
+ - String keyed command meta
 
 ## [1.2.0] - 2020-12-07